Will change PlayExpert to PlayerAgent class and add ChhooseMove method.
Decided that PlayExpert is basically deciding which moves to play, which gives it the role of a player (in contrast to the rulebook class which handles the rules and has the referee's role). So it should be the base class for different kind of player classes (AI, HMI, network interface).