Add project for GoogleTest framework (WIP)
Will convert most unit tests to GoogleTest framework to use GoogleMock
instead of dependencies or self-written mock classes.

Currently only builds with MSVC2017 32 bit compiler
(need to build Python libraries for 64 bit).
MinGW fails to build with Python 3.6 library, this might be fixed with
newer Python 3.7 and 3.8.
MSVC2015 fails to build with GoogleTest due to the compiler
not conforming to C++ standards.

But I plan to switch from Python to JavaScript as script language anyway.
The latter is supported by Qt5 via the QJSEngine class, which will reduce
a lot of hassle caused by the Python integration.